seniorcare_backend/requirements.txt
flask
firebase_admin
gunicorn
python-dotenv
fitbit
